% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ber_income}
\alias{ber_income}
\title{Income in Berlin, Germany}
\format{
A dataframe with 3,000 rows and 4 columns:
\describe{
\item{\code{id}}{Character vector containing fake IDs demonstrating what real
survey IDs could look like}
\item{\code{income}}{Ordered factor containing fake gross household income data
in groups of 500€ starting from 1000€ and ending at 5000€.}
\item{\code{x}}{Numeric vector containing the x coordinate in meters.}
\item{\code{y}}{Numeric vector containing the y coordinate in meters.}
}
}
\source{
Data is based on the results of the Census 2022.

© Statistische Ämter der Länder und des Bundes, 2024

Available under: \url{https://www.zensus2022.de/EN/Census_results/_inhalt.html}
}
\usage{
ber_income
}
\description{
A dataframe containing synthetic data on household income in Berlin. The
data contains coordinates projected in EPSG:3035.
}
\details{
The household distribution is based on dwelling data from the Census 2022.
Households can only be sampled in 100m INSPIRE grid cells with at least
one dwelling.

The income distribution is modelled based on Census 2022 data on average
cold rent, the share of owner occupiers, and the average living space per
inhabitant. Normally distributed noise was added to simulate realistic
spatial patterns.
}
\examples{
head(ber_income)
}
\keyword{datasets}
