% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nrw_chargers}
\alias{nrw_chargers}
\title{Charging stations in NRW, Germany}
\format{
A dataframe with 53 rows and 6 columns:
\describe{
\item{\code{ags}}{5-digit official municipality key (AGS). The first two
digits represent the state, the third represents the government district,
and the last two represent the district.}
\item{\code{count}}{Number of charging stations per district.}
\item{\code{fee}}{Share of charging stations that require a fee
for using the charging station.}
\item{\code{public}}{Share of public charging stations, i.e., stations
who provide unconditional access.}
\item{\code{capacity}}{Mean number of vehicles that can be charged at
the same time.}
\item{\code{operator}}{Most common charging station operator.}
}
}
\source{
© OpenStreetMap contributors, ODbL 1.0.
\url{https://www.openstreetmap.org/copyright}
}
\usage{
nrw_chargers
}
\description{
A dataframe containing crowdsourced data on electric charging
stations in North Rhine-Westphalia. The data is aggregated to the district
level and contains the respective municipality keys (AGS) for each district.
}
\details{
Charging station locations and details are taken from the OpenStreetMap (OSM)
database retrieved through the \href{https://postpass.geofabrik.de/}{Postpass}
API. A charging station is either a point or a polygon with the OSM tag
\code{amenity=charging_station}.
}
\examples{
head(nrw_chargers)
}
\keyword{datasets}
