% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_api.R
\name{sora_api}
\alias{sora_api}
\alias{sora_set_api}
\title{Get and set SoRa API}
\usage{
sora_api()

sora_set_api(url = NULL, verbose = NULL)
}
\arguments{
\item{url}{A valid URL to a SoRa API. If \code{NULL}, sets the URL to
the default public API of the SoRa service. Defaults to \code{NULL}.}

\item{verbose}{A logical, character string or \code{NULL} that specifies if
and how to produce messages. If \code{FALSE}, suppresses all informative
messages. If a path is provided, stores all messages in the specified file.
If \code{NULL}, verbosity depends on \code{getOption("sora_verbose")}. Use
the latter to set verbosity across the entire session, e.g. to suppress
messages from all package functions. Defaults to \code{NULL} and
\code{getOption("sora_verbose")} defaults to \code{TRUE}, i.e., display
messages where reasonable.}
}
\value{
Both functions return the currently (or newly) set SoRa API,
but \code{sora_set_api} returns it invisibly.
}
\description{
Set a non-default URL to communicate with the SoRa API. This is necessary
when working from a secure room or when connecting to a different SoRa
service (e.g., from a different country). By default, the \code{sora}
package sends requests to the default public API, which is not
available in secure rooms.

If you are working inside a secure room, make sure to ask the staff for
the SoRa URL. You can then set it using \code{sora_set_api()}. To retrieve the
currently set API, use \code{sora_api()}.
}
\details{
\code{sora_set_api} is only able to set an API URL for the current session.
It is reset upon starting a new session. To set a persistent path,
set the environment variable manually inside your system's environment
variable storage or use an \code{.Renviron} file.
}
\examples{
# Get the current API
sora_api()

# Set a new API, e.g., inside an air-gapped network
sora_set_api("ftp://secure-room-data/")
sora_set_api("https://10.1.1.1/")

# Get the new API
sora_api()

# Reset to the public API
sora_set_api()
}
