% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_objects.R
\name{sora_custom}
\alias{sora_custom}
\alias{sora_provided}
\alias{sora_admin}
\title{SoRa datasets}
\usage{
sora_custom(.data, crs = NULL)

sora_provided(.data, survey)

sora_admin(.data, schema)
}
\arguments{
\item{.data}{Coordinates or identifier to coordinates, depending on the
function.

\describe{
\item{\code{sora_custom()}}{A plain dataframe containing three columns
\code{id}, \code{x}, and \code{y} or an sf dataframe containing point
geometries. If an sf dataframe is provided, the geometry column is
converted to coordinates and the CRS is extracted automatically. If the
columns \code{id}, \code{x}, and \code{y} cannot be found in the data, the
first three columns are selected instead.}

\item{\code{sora_provided()}}{A character vector of IDs that distinctly
identify pairs of coordinates in the survey dataset provided in the
\code{survey} argument.}

\item{\code{sora_admin()}}{A dataframe containing two columns: one column
containing the respondent/case ID (\code{case_id}) and one column
containing the area identifier (\code{area_id}). An area identifier could,
for example, be a municipality key (AGS) or zip code. Note that these
columns can either be provided by name (\code{case_id} and \code{area_id})
or by position in the dataframe (positions 1 and 2).}
}}

\item{crs}{EPSG code of the coordinate reference system (CRS) of your data.
Can be one of the following: 3035, 3857, 4326, 4647, 25832, 25833. If not
provided and \code{.data} is of class \code{sf}, the EPSG code is extracted
automatically.}

\item{survey}{ID of a survey. You can look up possible values in the data
picker (see \code{\link{sora_datapicker}} or its web version).}

\item{schema}{Identifier of an administrative area schema. You can look
up possible values in the data picker (see \code{\link{sora_datapicker}}).
Example values include \code{"AGS2019"} and \code{"PLZ2024"}.}
}
\value{
List objects of classes \code{sora_custom}, \code{sora_provided},
or \code{sora_admin}, which can be used in \code{\link{sora_request}}.
}
\description{
Specify the data that you wish to link to the spatial context.
Generally, you have three ways to specify data to the SoRa service:

\itemize{
\item{Provide your own data using \code{sora_custom()}.}
\item{Specify a dataset provided by SoRa using \code{sora_provided()}.
Note that, in the public API, you can only access structural data of these
datasets. To work with the real data, you need to be inside a secure room.}
\item{Provide regional keys to link areal units to their spatial
context, e.g., zip code areas or municipalities, using \code{sora_admin()}.}
}

These objects can be used to feed linking jobs using
\code{\link{sora_request}}.
}
\examples{
sora_custom(ber_income, crs = 3035)
sora_provided(c("0001", "0002", "0003"), survey = "soep.core.v39")

admin <- data.frame(
  case_id = 1:3,
  area_id = c("05 3 15 000", "05 3 15 000", "05 3 15 000")
)
sora_admin(admin, schema = "AGS")
}
