% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_id.R
\name{sora_id}
\alias{sora_id}
\alias{sora_assemble_id}
\alias{sora_split_id}
\title{Split and assemble IDs}
\usage{
sora_assemble_id(.data, id_col = NULL, out_col = NULL, collapse = "_")

sora_split_id(.data, id_col = "id", out_col = NULL, sep = "_")
}
\arguments{
\item{.data}{A dataframe. The dataframe should contain the columns
specified in \code{id_col}.}

\item{id_col}{A character vector of columns in which the
necessary ID info is stored in.

\itemize{
\item{For \code{sora_assemble_id}, the columns from which to assemble
the ID. Defaults to \code{NULL}, or all columns.}
\item{For \code{sora_split_id}, a single column in which the ID strings
are stored. Defaults to \code{id}.}
}}

\item{out_col}{A character vector of column names that should be generated.

\itemize{
\item{For \code{sora_assemble_id}, a single column. Defaults to \code{id}.}
\item{For \code{sora_split_id}, a character vector with the same length
as there are ID components in \code{id_col}. Defaults to \code{NULL},
or \code{"id_*"} where \code{*} is the component number.}
}}

\item{collapse, sep}{A character string by which to combine or split the ID
components. Defaults to \code{"_"}.}
}
\value{
A dataframe with added columns. \code{sora_assemble_id} adds a
single column with the assembled ID strings, \code{sora_split_id} adds a
number of columns equal to the number of ID components in \code{id_col}.
}
\description{
Helper functions to handle ID columns in the SoRa input/output.

\itemize{
\item{\code{sora_assemble_id}: Combines multiple columns to a single ID
column.}
\item{\code{sora_split_id}: Splits an ID column to multiple separate
columns.}
}

These functions are particularly handy (and were designed) to easily handle
the IDs of SOEP datasets. Use \code{sora_assemble_id} to generate ad-hoc
IDs from multiple dataset columns.
}
\details{
\code{sora_split_id} can generate \code{NA} values when single ID strings
lack components that are present in other IDs. For example
\code{c("id_a", "id")} would result in
\code{data.frame(id_1 = c("id", "id"), id_2 = c("a", NA))}. Similarly,
\code{sora_assemble_id} skips \code{NA} values, i.e., \code{c(id, NA, "a")}
becomes \code{"id_a"}.
}
\examples{
ids <- tibble::tibble(id_1 = "id", id_2 = c("a", "b", NA), value = c(1, 2, 3))
ids

assembled <- sora_assemble_id(ids, id_col = c("id_1", "id_2"))
assembled

split <- sora_split_id(assembled)
split

identical(ids, split)
}
