% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_problems.R
\name{sora_problems}
\alias{sora_problems}
\title{Linking problems}
\usage{
sora_problems(out)
}
\arguments{
\item{out}{A dataframe of class \code{\link{sora_results}} that produced
row-level error messages.}
}
\value{
A dataframe containing those IDs that produced row-level error
messages and a column \code{message} that contains their error messages.
}
\description{
Sometimes linking jobs do not fail but cannot compute linking results
for each row of the input data. In these cases, \code{\link{sora_results}}
informs you about which rows are erroneous. To see \emph{why} they failed
you can use \code{sora_problems}. It returns a dataframe containing the
problematic rows and the error message they produced in the SoRa API.

Given a results dataframe \code{out}, you can remove these cases from
the dataset by typing:

\preformatted{out[!is.na(out$count), ]}
}
\examples{
\dontshow{if (isTRUE(Sys.getenv("SORA_RUN_EXAMPLES")) && sora_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Create a problematic dataset
sora_data <- sora_custom(ber_income, crs = 3035)
sora_data$`coordinate-items`$x[4] <- NA

# Construct and send a request
spat_data <- sora_spatial("ioer-monitor-s02rg-2023-1000m")
linking <- sora_linking(
  method = "aggregate_attribute",
  selection_area = "square",
  length = 2500,
  output = c("mean", "median", "sd")
)

out <- sora_request(
  sora_data,
  link_to = spat_data,
  method = linking,
  wait = TRUE
)

# Check for errors
problems(out)
\dontshow{\}) # examplesIf}
}
