% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_objects.R
\name{sora_provenance}
\alias{sora_provenance}
\title{Provenance}
\usage{
sora_provenance(
  visible = FALSE,
  name = NULL,
  pid = NULL,
  url = NULL,
  description = NULL,
  authors = NULL
)
}
\arguments{
\item{visible}{Whether to generate provenance record or not. If
\code{TRUE} produces a provenance record according to the input to this
function. Defaults to \code{FALSE}.}

\item{name, pid, url}{When using a custom dataset using
\code{\link{sora_custom}}, these arguments specify the name, ID, description,
and URL of the dataset. For optimal reproducibility, make sure to provide
accurate information to these arguments.}

\item{description}{A character string describing the subsample analyzed
in the linking job. For optimal reproducibility, make sure to provide
accurate information to these arguments.}

\item{authors}{Name of the author for the request. Will be used for the provenance record.}
}
\value{
An object of class \code{sora_provenance}.
}
\description{
Configure if and how a provenance record for a linking job
should be generated. This function is mainly used to be used as an
argument to \code{\link{sora_request}}. You can configure the visibility
of the provenance record. When providing custom coordinates, you can also
provide the dataset's metadata if applicable.
}
\examples{
# The default, `FALSE`, does not produce a record
sora_provenance()

# Set to `TRUE` to produce a reproducible report for your linking job
sora_provenance(TRUE)

# When providing custom data, you can also use `sora_provenance()`
# to provide necessary information to create a record
data <- sora_custom(ber_income, crs = 3035)
sora_provenance(
  visible = TRUE,
  name = "Household income, Berlin",
  url = "https://sora-service.org/",
  authors = "John Doe"
)
}
