% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_objects.R
\name{sora_spatial}
\alias{sora_spatial}
\title{Geospatial dataset}
\usage{
sora_spatial(id = NULL, .data = NULL, ...)
}
\arguments{
\item{id}{An ID of a spatial dataset provided by SoRa. See
\code{\link{sora_datapicker}} or its web equivalent to see which
IDs are available. If you want to provide your own spatial data, see the
\code{.data} argument.}

\item{.data}{A plain dataframe containing two columns \code{x} and \code{y}
or an sf dataframe containing any geometry type. Use this argument if you
intend to provide custom spatial data. If you want to use spatial data
provided by the SoRa service, see the \code{id} argument.}

\item{...}{Further arguments passed to the API. See
\code{\link{sora_datapicker}} for a list of available arguments.}
}
\value{
An object of class \code{sora_spatial}.
}
\description{
Specify the the geospatial dataset to be used in the
\code{link_to} argument of \code{\link{sora_request}}.
}
\examples{
# When providing an ID you tell SoRa to select a pre-made dataset
# from the service provider
sora_spatial("ioer-monitor-s02rg-2023-1000m")
}
