% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_request.R
\name{sora_wait}
\alias{sora_wait}
\title{Waiter}
\usage{
sora_wait(job_id, verbose = NULL, api_key = NULL)
}
\arguments{
\item{job_id}{A job ID as returned by \code{\link{sora_request}}.}

\item{verbose}{A logical, character string or \code{NULL} that specifies if
and how to produce messages. If \code{FALSE}, suppresses all informative
messages. If a path is provided, stores all messages in the specified file.
If \code{NULL}, verbosity depends on \code{getOption("sora_verbose")}. Use
the latter to set verbosity across the entire session, e.g. to suppress
messages from all package functions. Defaults to \code{NULL} and
\code{getOption("sora_verbose")} defaults to \code{TRUE}, i.e., display
messages where reasonable.}

\item{api_key}{A valid SoRa API key. Required to communicate with the SoRa
API. You can request a key by filling out
\href{https://sora.gesis.org/unofficial/sora-user-mod/users/request-api-key}{this form}
on the SoRa website.

If \code{NULL}, reads a key from the \code{SORA_API_KEY}
environment variable.

If "ask" (and the session is interactive), opens a masked prompt where you
can enter your key. The entered key is persistently stored in an environment
variable and can be used across sessions.

Please note that you are advised not to enter your key directly using this
argument. If you do, please make sure that you do not publish any code files
containing the key in plain text. Failure to do so can compromise your
access to the SoRa API.}
}
\value{
The same as \code{\link{sora_results}}, a dataframe containing
the case IDs and all requested output statistics.
}
\description{
Helper function that takes a job ID and waits until the linking job behind
it has finished, then automatically fetches the results. This function is
used for the \code{wait} argument in \code{\link{sora_request}}. Starting a
waiter can be particularly useful in continuous workflows where the
next steps depend on the availability of the linking results.
}
\details{
By default, \code{sora_wait} waits for one second before polling the
job status again. You can change the option \code{sora_wait_delay} to
poll more or less frequently than that.
}
\examples{
\dontshow{if (isTRUE(Sys.getenv("SORA_RUN_EXAMPLES")) && sora_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
job <- sora_request(
  sora_custom(ber_income, crs = 3035),
  link_to = sora_spatial("ioer-monitor-s02rg-2023-1000m"),
  method = sora_linking(
    method = "aggregate_attribute",
    selection_area = "square",
    length = 2500,
    output = c("mean", "median", "sd")
  )
)

sora_wait(job)
\dontshow{\}) # examplesIf}
}
