% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_datapicker.R
\name{sora_datapicker}
\alias{sora_datapicker}
\alias{sora_arguments}
\alias{sora_dp_overview}
\alias{sora_dp_get_id}
\title{Data picker}
\usage{
sora_datapicker(content = c("provided", "admin", "spatial", "linking"))

sora_arguments(method, provider = "IOER")

sora_dp_overview(
  data_dp = NULL,
  sort_by = "Raster",
  content_dp = "spatial",
  arg_1 = "title",
  arg_2 = "spatial_resolution"
)

sora_dp_get_id(
  data_dp = NULL,
  indicator,
  dataset_id = "dataset_id",
  content_dp = "spatial"
)
}
\arguments{
\item{content}{The type of dataset for which to retrieve details. The
following values are allowed:

\itemize{
\item{\code{"provided"}: Metadata of coordinate datasets provided by SoRa
through \code{\link{sora_provided}}}
\item{\code{"admin"}: Metadata of administrative area datasets recognized
by SoRa through \code{\link{sora_admin}}}
\item{\code{"spatial"}: Metadata of geospatial datasets provided by
SoRa through \code{\link{sora_spatial}}}
\item{\code{"linking"}: Metadata of available linking methods that can
be used in \code{\link{sora_linking}}}
}}

\item{method}{A linking method that is used in \code{\link{sora_request}}
and \code{\link{sora_linking}} to link a (survey) dataset to a
spatial dataset. \code{sora_arguments} returns the arguments related to
the specified linking method.}

\item{provider}{ID of a linking service provider to perform the
spatial linking. Defaults to \code{"IOER"}.}

\item{data_dp}{datapicker}

\item{sort_by}{if arg_2 is 'spatial_resolution' the data can be sort by the
'Raster'. This value is the default.}

\item{content_dp}{selected table like "spatial"}

\item{arg_1}{argument 1. The default is 'title'.}

\item{arg_2}{argument 2. The default is 'spatial_resolution'.}

\item{indicator}{is the 'title' output in the 'datapicker'}

\item{dataset_id}{argument for the 'dataset_id' can be depends of the dataprovidor.}
}
\value{
\itemize{
\item{\code{sora_datapicker} returns a dataframe containing
details about provided coordinate datasets, administrative areas,
geospatial datasets, or linking methods depending on the value passed to
\code{content}. Some common columns include:

\itemize{
\item{\code{provider}: The provider of the dataset.}
\item{\code{dataset_id}: Dataset ID that is used in
\code{\link{sora_provided}} and \code{\link{sora_spatial}}.}
\item{\code{description}: Description of the dataset. Used in the
provenance record.}
\item{\code{pid}: Persistent identifier of a dataset, usually a DOI.}
\item{\code{required_permissions}: Information about the terms of use
required to access the data. These need to be accepted when registering
for an API key.}
}}
\item{\code{sora_arguments} returns a dataframe where each row is a
argument that can be passed to the method specified in \code{method}
and each column gives further details about the argument.}
}
}
\description{
Download and read the information from the SoRa data picker.

This function provides a programmatic interface to the web version of the
data picker, which can be found under
\url{https://sora.gesis.org/unofficial/datapicker/}.

Get an overview of the datapicker provided by 'SoRa'.
The output is a crosstable of the both arguments ('arg_1' and 'arg_2').

get an overview of the possible dataset id which can be selected.
The output has this columns: 'title', 'time_frame', 'spatial_resolution' and 'dataset_id'
}
\examples{
\dontshow{if (isTRUE(Sys.getenv("SORA_RUN_EXAMPLES"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sora_datapicker("admin")

sora_datapicker("spatial")

sora_arguments("nearest_poi")
\dontshow{\}) # examplesIf}
}
